services/ourcents/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.6.1
pydantic-settings>=2.8.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
openai>=1.12.0
google-generativeai>=0.3.2
sqlalchemy>=2.0.27
bcrypt>=4.1.2
pillow>=10.2.0
python-dateutil>=2.8.2
pytz>=2024.1
